fix: update to ditto-kotlin 5.0.0-rc.1
execute() now returns Unit; use executeRaw() where results are needed.
DittoDiskUsageItem.sizeInBytes changed from Int to Long.

+++ DittoToolsAndroid/src/main/java/live/ditto/tools/diskusage/DiskUsage.kt
@@ -37,6 +37,6 @@ fun DiskUsageScreen(
 
 data class DiskUsage(
     val relativePath: String = "ditto/ditto_store",
-    val sizeInBytes: Int = 0,
+    val sizeInBytes: Long = 0L,
     val size: String = "Calculating...",
 )
